Windows gs console window unicode changes.
Changes to dwtext.c to allow unicode operation.

The window is created as a Unicode one, so we get unicode character codes
delivered to us as they are typed. We convert these incoming key codes
into utf8, as the gs core expects. Any output from the core is converted
from utf8 to unicode, and the window now displays unicode.

The net effect is that the window looks like it's behaving exactly like
a unix window does.

The copy and paste functions are updated to cut and paste unicode too.

As usual, building with WINDOWS_NO_UNICODE preserves existing behaviour
(useful for windows 95/98/me builds).

The sole difference in behaviour is that the copy function now removes
trailing blank lines from the output, as this is never what you want.

CLUSTER_UNTESTED as windows only changes.