Bug 692263 spot and elide glyphs with zero dimensions.
A glyph with a zero scale in either dimension could cause divide by
zero errors in a number of places in the FAPI *interface* code
(that is: fapi_ft.c, fapiufst.c and fapibstm.c). So catch and handle
the case early by skipping such case. The wrinkle is charpath operations
which *must* create a path, even of zero dimensions.

No cluster differences.