More unicode changes for gswin32
Taken on (most of) a patch from Russell Lang to dwtext.c, with bugfixes
to the changes I made yesterday. The bits of the patch I didn't take
on, were to do with running gswin32 as a true Unicode application.

Currently we are an ANSI application with a Unicode window. Moving to
be a true Unicode application may have knock on effects I am unsure of
(and would certainly leave us different to gswin32c.exe). I don't think
we lose anything by staying as we are.

CLUSTER_UNTESTED.