Fix for aliasing of icc link hash code
This fixes a problem that existed in the computation of the ICC
hash code.  This was the source of the problem in Bug 692265 (bug fixed with
this commit in file gsicc_cache.c).  Also  some error checking code and
addition of code to handle the transfer function when doing the fast
thresholding of color images.