Add new 'plan' family of planar devices.
Equivalent to the plib devices, but without the necessity to always band,
and not using the band donor.

From the end users point of view, these output ppm etc and are probably
therefore not that useful as is. They serve the twin purposes of allowing
us to test planar output, and to act as a basis for devices that need
planar non-interlaced output.