Second attempt to add plan devices.
Add new 'plan' family of planar devices.

Equivalent to the plib devices, but without the necessity to always band,
and not using the band donor.

From the end users point of view, these output ppm etc and are probably
therefore not that useful as is. They serve the twin purposes of allowing
us to test planar output, and to act as a basis for devices that need
planar non-interlaced output.

The development of these devices showed a problem with
gx_get_bits_return_pointer in the planar case. When called, it would
assume that the start of the first line of the first plane was at
'base', and that subsequent planes could be reached by adding
raster * height to it. Unfortunately, clist_rasterize_lines resets the
buffer device at the end, so that height is unrelated (often 1).

The fix implemented here is to change gx_get_bits_return_pointer to take
a pointer to the line pointers, rather than a direct base value. This
means we can always find the subsequent planes correctly.

In my initial attempt at this, I broke the code by failing to notice that
gdevmem.c passes the same base pointer to gx_get_bits_copy. Fixed here.
Cluster testing shows it be OK this time around.