Bug 689546: Add clist.dev to the core lib dependencies.
I previously noted that the clist device is essentially always required
now, but I used a less than ideal method to address it: having it as
a dependency of the gdevprn.o object.

The correct solution, done here, is to include the clist device in the
libs.dev file, thus making it a direct requirement of the core lib.

No cluster differences.