New lib.mak dependency checker, and update lib.mak using it.
Invoke gs/toolbin/checkdeps.pl and it performs rudimentary
and easily confused checks on the dependencies in lib.mak.
It may be simple, but it spots MANY problems, the important
of which are fixed here.

It gets easily confused by files not being in the expected
paths (but these are easy for an operator to spot and ignore).

No cluster differences expected.