Move planar dev_spec_op into core planar device.
Rather than having plib and plan devices provide a dev_spec_op that
responds to gxdso_is_native_planar, put the responsibility for this
into the core planar memory device.

This not only simplifies the code, it actually fixes a problem whereby
the memory device was overwriting it with the default.