Fix for Bug 692286 - Image too light -- not fully opaque
when I did some of the work on the compositor queue I had set things up so that the
update of the blend parameters would be written only into the same bands as the group
push. The approach of writing only in the bands of the group bounding box will not
work in general, for example if we have nested groups that only partially overlap. The
solution is that we have to write the blend parameters in all bands.