ps2write output crashes with limitcheck error after converting large documents
Adopting patch from 'Steve166' (bug # 692172) with a little gratuitous
reformatting. (note that opdfread.ps is now compiled in as opdfread.h)

This converts the directory of objects from an array to a dictionary of dictionaries
which allows for a more or less unlimited (64k*64k) number of objects.

No cluster differences expected.