Permit the calculation of CIDSet and CIDToGIDMap size, for PDF/A validation
The SolidWorks PDF/A validator complains about the CIDSet we produce for
embedded TrueType fonts, converted to CIDFonts, when creating PDF/A files.

I'm still not exactly sure what its complaining about but I'm working with
our customer to resolve this. In the meantime, this code makes it easier
to play with teh emission of CIDSet and CIDToGIDMap.