Bug #692218 Create comatible CIDSet
The fix in Bug 692218 did not (apparently) work for the SolidWorks
PDF/A validator. This patch sets the number of entries in a CIDSet to
be the number of glyphs in the font program, and makes the number of entries
in a CIDToGIDMap the same.

Together with the commit 27b740 here:

http://ghostscript.com/pipermail/gs-cvs/2011-June/013024.html

this now seems to work correctly.