Bug 692274: handle vastly different magnitudes in x/y scale
If the difference in the scale on the x and y axes is large
(in this case, a factor of 1000), the value scaling we do
to pass into Freetype can run over/under flow when applied
to both. The solution is to apply the value scaling
separately for x and y axes. But the quality is very
marginally better when the two are scaled together, so we
do that normally, and only fall back to the separate handling
when the x and y scales differ by more than 512.

No cluster differences expected.