Add copy_plane to rop source device.
This uses 'creative engineering' (aka a blatant hack) to allow copy_color
to be implemented in the "rop source" device.

We implement a copy_plane entrypoint for the rop source device, that
does exactly the same as copy_color, except that it abuses the lop
value passed to the gx_device_color_fill_rectangle by setting a new
bit (lop_planar) and pickling the plane number into the high bits.

This is detected in mem_planar_strip_copy_rop and unpicked to a call
to the appropriate plane.

Care must be taken when doing rops on the planar device, as
processing any rop on the device that involves 'D' calls get_bit_rectangle
which can cause a call to convert from planar to chunky if mishandled.
The solution is to ensure that we always remove our get_bits_rectangle
implementation when passing on a ropping call.