Update to libjpeg 8c.
Combines (from libs-update branch) the commits:

Commit a vanilla copy of libjpeg-8c.

And the makefile changes to let it build.

Commit our patch to the libjpeg source.

Add our patches to the jpeg-8c src.

This changes every cluster test which contains DCT/JPEG data, but
they are all very small colour shifts that are invisible
to the naked eye - as expected by this type of update.