bug #692304 Don't emit '0 g' for 'erasepage'
showpage causes a fill of an empty path, apparently as part of the erasepage mechanism.
When used with pdfwrite, with -dUseCIEColor -dPDFA and no Output Intent specified
this can lead to us emitting a '0 g' in a PDF/A file, whic his only valid if an OutputIntent is specified.

Fixed by ignoring empty paths. Despite the comment in gdev_pdf_fill_path we
do not seem to need to do this for text or clips. Probably since I moved the
'pdf_set_drawing_color' code out of pdf_reset_color so that we could write colours in a text context.

This exhibits one progression, with ps2write 'Bug6901014_CImg_flyer.pdf' now
draws all the output instead of dropping portions of it due to a 'nocurrentpoint' error.