More 'aux' makefile changes.
Configured gs builds were still not putting the aux directory in the
right place. Reworked here.

This has meant changing the horrible 'RELDIR' hack for a slightly
less horrible hack using 'BUILDDIRPREFIX'. In turn this has meant
that the shared object targets have one mroe layer of recursion in
the make, but it's still at the top level, and (arguably) clearer
now than it was before.