Tweak to gsicc_manage.c to avoid potential leak.
Examining the code, if gsicc_set_icc_directory is called with namelen=0
then the existing code will allocate a 1 byte buffer and store it in
the icc_manager.

If the same function is called again, in the same way, this buffer
will not be freed before it is overwritten.

The fix is simply to test for icc_manager->profiledir being non NULL
rather than the namelen > 0.