Implement gs_setcachesize properly.
The previous implementation simply set the cache size state value but
didn't attempt to rebuild the cache.  The new font cache size is
provided by size operand of setcacheparams.

We also limit the size operand to the values discovered used by CPSI, the
range is device dependent and the CPSI range appears reasonable.