Fix clist planar device detection.
When the clist interrogates a device to see if it is a planar one, it was
getting the wrong answer due to only the planar memory buffer device
responding to the question. We fix the main planar device to respond too
here.

Also, any device that doesn't understand the is_native_planar question
would respond -1; the existing code would take this to mean that it is
planar. The code now only triggers on positive responses.

No cluster differences expected.