Bug 692318: Ensure that compiler flags are used for the "aux" files.
For a normal host build, the build tools (genarch, genconf etc) should be
built with the same compiler flags as Ghostscript/Ghost*. In this case
the integer type used for encoded color values was not getting used
when compiling genarch.

Also, add a warning when tiffsep does have to skip one or more plates, with
a pointer to the relevant documentation. Lastly, update the doc to reflect
that the contone preview output may not be as expected if the job uses
overprint.

No cluster differences expected.