Shailesh Mistry's fix for 692311 - corrupt character rendering.
In adaptive compression mode the seed row for delta row compression
was not properly cleared.  With this change the code is now compatible
with HP monochrome printers, previously the Artifex code was
compatible with HP Color Laserjets.  We do not implement a device
specific switch - for example the new code could be enabled if the
current mode of emulation was PCL5E (mono PCL) and not PCL5C (color
PCL) allowing emulation of both product types.  However the corrupt
character output of HP color printers is probably never an intended
result.