Bug 688528: set ICC profiles for interal uses of bbox device.
If we create a bbox device internally (i.e. not as an output device)
we must still ensure that the device has ICC profiles set in it. In
this case, it simply inherits the profile array from the parent
device.

This also meant moving the ICC manager initialization to before the
device "open" call in gs_setdevice_no_erase() because the device set
can also open another device (in this case, pswrite's "open" has the
effect, via the vector device, of setting a bbox device instance, so
there must be an array of ICC profiles at that stage for bbox to
inherit.

No cluster differences expected.