Fix Bug 692331: SEGV in image interpolation.
In fixing bug 692225, it seems I broke this.

In the setup code that decides whether decoding is required or not, I
was incorrectly choosing to use the 'decode' case, when in fact the
'decoding' is done later for CIE colorspaces. As such I should have used
the no-decode case.

Fixing this appears to cure the problem.