Fix initialization in gs_nulldevice (PS nulldevice operator) Bug 692334.
The gs_nulldevice was setting the get_profile proc to gx_forward_get_profile so the
icc_array was coming from the current device in the graphics state but that did not
match the color_info num_components and polarity leading to incorrect values returned
if gscms_transform_color_buffer set the OutputFormat to 1 (for the nulldevice) of a
link handle, but a subsequent call to gscms_transform_color with the same link handle
expected more than one component.

The initialization of the default profile can only be done if the io_device_table has
been initialized since this function is called during gs_lib_init1 with the table NULL.

This change fixes 12-07D.ps differences introduced in the 'tagfix' patch, but does NOT
introduce any changes to gscms_transform_color_buffer to avoid changing the number
of components or issue a warning as mentioned in the bug.

EXPECTED_DIFFERENCES:

12-07D should revert to the output prior to the tagfix change (06df93f)