Bug 692338: address performance regression in FAPI
Previously I changed FAPI to only dispose of the outline or
bitmap part of the Freetype glyph, intead of the entire
internal glyph object, so as avoid freeing, re-allocating
and initialising a glyph object for every glyph. That seemed
to cause a memory leak, so I reverted it.

This commit does it properly, addresing the memory leak by
correctly freeing the memory that Freetype returns to us
*containing* the pointer to the bitmap or outline, as well
as the bitmap/outline data.

This is measurably quicker than the previous method.

No cluster differences expected.