Revise error/warning on PDF missing CIDFont.
Now that GS will attempt to continue interpreting a PDF after an error
in a given content stream, revise the message we emit when he hit a
missing CIDFont in a PDF condition to reflect that. If PDFSTOPONERROR
keep the previous message.

Also, add a "flush" at the end of each message to ensure they don't get
mangled by later, high priority messages.

No cluster differences.