Fix pattern cache tile accounting.
When putting tiles into the pattern cache, we estimate the size we need
up front (or in the clist case, calculate it exactly). We then clear
that many bytes in the pattern cache, and add the new tile using that
estimated size. When we come to evict we recalculate the exact size and
evict based on that.

This causes a mismatch whereby the pattern cache thinks it has the wrong
number of bits cached. When this is an underestimation, we just end up
using more memory than we'd like. When it overestimates however, we can
get into an infinite loop trying to evict things.

The fix is to record the number of bits we've accounted for in every
tile and thus always count the same coming out as we did coming in.

If we really want to account for the correct values rather than the
estimate we can do that, but we must adjust pcache->bits_cache and
ctile->used appropriately.

No bug associated with this, but it was triggered by a forthcoming
commit to make the pattern accumulator use planar buffers for planar
devices.

No differences in cluster testing.