Pattern accumulator changes to support planar output devices.
When creating the memory buffer for a painttype 1 pattern tile, check
to see if the underlying device is planar - if it is, then create the
memory buffer as planar too.

The code to tile out the planar pattern buffers still needs fixing, but
this allows halftoning jobs to run to completion at least.

No cluster differences shown.