Fix for missing ICCProfilesDir during some device init. Move profiledir to gs_lib_ctx.
The business of trying to keep the profile directory in the device and keeping it in
sync with the value used in the icc_manager led to problems when some devices were
initialized. This change moves the profiledir (and its len) into the gs_lib_ctx_t
(pseudo- global) which is accessible by the icc_manager and devices via the
gs_memory_t pointer.

Remove the no longer needed functions to set and sync the profile in the device and
replace gsicc_set_icc_directory with gs_lib_ctx_set_icc_directory.

Thanks to Michael Vrhel for adding to this patch to make it build on linux and
build the other parsers.