Add finalize function to pattern clist device.
This may need to do more than it does, but the immediate drive for
this is to correctly drop the reference count on the target device
on closedown. This is sufficient to stop the SEGV seen on the final
garbage collection done on stars.pdf.

The garbage collector was finding an unfreed pdf14 device, and stepping
into the target onto to find that the target had been destroyed already.

No cluster differences shown in testing.