Fix Bug 692355; gswin32 has garbage chars in window title bars.
Existing ghostscript includes windows.h without defining UNICODE, but
calls the unicode versions of functions where appropriate. Here we
move to defining UNICODE before including windows.h, but we keep to
the practise of calling A or W specific variants as much as possible.

Partly this is because of time before the release, but mostly a fear
that this might lead to a Pandoras box of changes.