Fix bug 692235: Rendering to 8bit colors shows bad colors.
The display device uses an RGBK space for 8 bit rendering. Currently
this corresponds to a palette of 96 colors; 2 bits each for R,G,B,
supplemented by 16 greys.

This review alters it to use a 6x6x6 RGB cube, supplemented by 40 grey
levels.

In fact, this is slightly wasteful as we repeat the black and white
representations as color and as greyscale, but it's a huge improvement
on what we had before.

No cluster differences expected as we don't test the display device.