Fix for bug 692364  in which we were were not handing the initial DeviceGray profiles correctly
When the graphic state is started, the color spaces are first set to DeviceGray.  In certain documents
this is then the color space that is used for drawing and in fact the erase page will initially
occur with this color space.  If I have a destination ICC profile like CIELAB, then this is
a problem since we will end up mapping DeviceGray to RGB with the old procedures giving us a
CIELAB value of [255 255 255] which is a very bright red.  With this fix we in the
DeviceGray mapping we fix the colorspace to be ICC based and pass then handle the procedure with
the ICC mapping.    A minor fix was also needed in the pdf14device put_image where we use a
DeviceGray color space, which happens to have an ICC profile.  We now correctly create the
ICC profile.  Note that this will cause significant diffs in the cluster push as DeviceGray
colors that were mapped to pure K before for a CMYK device (due to the fact that we were not
installing the proper color space) will now map to composite black.  This
issue will be addressed shortly by another commit where we by default map all DeviceGray
colors to K only in the CMYK devices.