Backout commit 7f5d3d7; 8 bit display device palette change.
Reverting earlier commit made to attempt to fix bug 692235. While
this change does give the display device a much nicer palette
to work with in 8 bits, it will break with all code that currently
exists to interface to the display device, such as gsview, as the
palette is not exported with the data.

Due to the looming release and the questions remaining to be answered
on the bug, we pull the change for now.