Add functionality to the textwrite device.
The existing textwrite device was a null device which emitted nothing. This
version actually extracts text from input documents. It uses ToUnicode CMaps,
GlyphNames2Unicode tables and glyph names in an attempt to map incoming text into
Unicode (UTF-16) values.

It currently has one switch 'TextFormat' which has two possible values, 0 or 1.
-dTextFormat=0 outputs a 'layout' representation of the original text. This
will attempt to produce text output which is similar in appearance to the original.

-dTextFormat=1 simply emits all the text as it is encountered, along with information
about the text state. It outputs the Font name, point size, rendering mode and
writing mode (WMode) as well as the start and ending co-ordinates of the text on the page
and the Unicode string representing the text. This is intended for those who
would like to manipulate the text into other formats.

The Adobe Glyph Names list is now embedded in the executable twice, in two
incompatible forms. A future enhanvement will be to only embed the list once.