Further updates to mkromfs ps compaction.
Refactor the postscript compaction routines into a state machine.
Add options to compact using binary postscript encodings, currently
disabled until we drop language level 1 from the interpreter.

Further refactor it to allow it to work from a generic file-like
interface, rather than a file. This enables us to use it on the
merged file (which is held in memory, not on disc) too.

With binary encoding enabled we get the rom down to 6.8Meg. Without
that we get 7.1Meg. Without any compaction, 7.8Meg.

No cluster differences expected.