Fixed the problem that PCL-XL output files got huge after introduction of color management
The pxlmono and pxlcolor output devices produce PCL XL (PCL 6). With
the introduction of color management, the size of the output files
often got 10 times larger as before and so PCL XL jobs often did not
print as they exhausted the memory of the printer.

According to the contributor of this patch, Shailesh Mistry, the PDFs
leading to that problem have a DeviceRGB image which is translated
into an ICC profile and passed up but the pcl/pxl can not handle
it. This patch looks for incoming ICC profiles and resolves it into
the base colour which can be handled at this level.

The patch returns the size of the PCL XL output files to the old
values from Ghostscript 8.x.

Fixes bug 692329.