Fix for Bug 690137
The attempts to get the range properly set up for when we have an ICC profile for CIELAB.
The issue is that we may have vector drawings that are encoded across the range of real
CIELAB values like [100,0] [-128, 127] [-128, 127].  Image data of course is not encoded
in this form.  This fix corrects for this difference.