Add backtrace to memento SEGV handler.
Simple backtrace handling; thanks to Tor for the pointer to the
stackoverflow article.

This only works with libc based compilers, and may possibly get the
bottom couple of entries on the stack wrong. It also doesn't give
file and line references, but it does at least give the function
names.

Other suggestions in the article offered other ways of working, but
they all rely on other libraries, or non-portable features, so this
seems the best possible first step.