Bug 692367: add gs_memory_t arg to finalize method
By adding a gs_memory_t argument to the object "finalize" method, we can
dispense with the pointer-pun hackery that stores the memory context
in an extra struct array element of the IO device table, so it's availabe
in the finalize method.

Although primarily addresses one hack, this commit touches a number of files
because it affects every object with a "finalize" method.

This also addresses an error condition cleanup of a partially create IO
device table.

No cluster differences.