Suppress some warnings in gdevcups.c
As cups_set_color_info() now does things that can fail, it has to return
an error value (getting rid of the "code set but never used" warning).

Also, add a "default:" case to a switch to get rid of the warning about
enum values not handled in a switch - the default case does nothing, so
behaviour is unchanged.

Finally, tidy up the initialization of an array of arrays of arrays, getting
rid of a warning about missing braces.

No cluster differences.