Fix bug 692372: Add finalize for imager_state to ref_count decrement icc structs.
The graphics atates were being freed by the 'alloc_restore_all' done during gs_lib_finit
but the icc_link_cache was not getting its ref_count decremented, so it never freed the
semaphore which caused handles to be lost by windows. The rc_gsicc_profile_cache_free
function is also fixed to save the 'next' pointer to avoid dereferencing freed memory.

Note that we _should_ be able to call gs_imager_state_release to decrement the counts
of ALL of the elements, but the ref counting of the other elements is _so_ badly
maintained that some elements (dev_ht and halftone) are prematurely freed by this.
This change is enough to fix the bug and pass cluster regression.