Fix pattern matrix problems when the pattern is used in a Form and creating PDF.
Bug #691800 exhibited a problem using a Shading (type 2 pattern) in a form.
This is unusual because hte only way to get a Form out of pdfwrite is to involve transparency.

When a pattern is used in a form, we do not apply the current CTM, but the
CTM of the enclosing context. In this case we were still applying the outermost
context, not the CTM of the page stream (which is the enclosing context of the form).

Fixing this revealed other problems.

I've moved the form counting out of the transparency group and into the form
dictionary creation code where it makes much more sense. This allows me to
change the 'FormDepth' checks so that we only undo the page content scaling
if the formdepth is 0.

In the long run we may need to keep a stack of gstatas (or more accurately the
CTM) so that if we get nested forms we can apply the correct CTM to any patterns
in the form. For now, this isn't required, we don't have any test files which nest
more than 1 form deep.

This fixes Bug #691800. I also see progressions in the following files:

Bug6901014_CityMap-evince-pdftopdf.pdf
Bug691783.pdf
fts_04_0403.pdf
fts_06_0608.pdf
fts_09_0919.pdf
fts_12_1202.pdf
fts_12_1205.pdf
fts_14_1416.pdf
fts_15_1508.pdf
fts_15_1510.pdf
fts_15_1512.pdf
fts_15_1514.pdf
fts_15_1516.pdf
fts_15_1518.pdf
fts_15_1520.pdf
fts_28_2803.pdf
fts_28_2804.pdf
fts_31_3107.pdf