Bug #692105. Increase the maximum input/output dimesnions of sampled functions.
The maximum number of inputs and outputs allowed in a sampled function is a
compile-time value. This particular PDF file had 20 colorants, and a sampled
function with 20 inputs and 4 outputs to convert it into CMYK.

This broke the maximu, even though GS had been compiled with GS_CLIENT_COLOR_MAX_XOMPONENTS
increased to allow the PostScript version (which doesn't use a sampled function)
to complete.

This patch permits a new compile-time define GS_CLIENT_SAMPLED_FN_MAX_COMPONENTS
which can be used to increase this value. If this is not declared then we now use
the GS_CLIENT_COLOR_MAX_COMPONENTS as it is likely we will encounter sampled functiosn
with that many components

No differences expected