Bug #692344 Make sure we do not dereference a NULL clip path when freeing
When creating the 'privately allocated' parts of a gstate it is possible to
fail to allocate a clip path. In this case we free the clip path, but if its
notbeen allocated then it is NULL, and the free routine unconditionally
de-references it.

Added a NULL pointer to check to prevent this. I *believe* this fixes the bug
but just changing the code made the allocation succeed when I tested it, so I
am not absolutely certain. This patch does fix a genuine failure though.