Bug 692435: fix PDF pre-rendering spot color search.
Before we start rendering a page, we search the resources for the page
(including its descendent objects), so that we can pass the information
to the device prior to rendering. What we weren't doing was searching
the parents of the page object (we were, however, doing so during the
stream interpetation for rendering), thus in rare, but valid files, we
could miss setting up the page device for spot colors.

No cluster differences expected.