Change planar devices to use encode/decode_color only.
Simplify the code by removing deprecated color mapping functions. In
particular this fixes a problem whereby encode_color for the plank
device was returning a color index encoded as rgb, and hence confusing
the overprint device.

CLUSTER_UNTESTED as the plan* and plib* devices are not tested at the moment.