Bug #692457 If a PCL pattern turns out to be a flat fill, still update the saved colour.
If a PCL pattern turned out to be a simple colour (black or white) we optimise
the output and don't emit a pattern. However, we do still need to update the
saved colour in our private state, because we will still be changing the colour. If
we don't do that then our saved state does not reflect the actual content of
the PDF file, and we can emit incorrect colours.

No differences expected.