Bug 692481; partial fix - jpeg data with excess 0xFFs.
Ghostscript passes buffers full of data to the jpeg library. The jpeg
library decodes one (or more) scanlines of data from this data and
returns it to ghostscript.

In the event that the jpeg library can't decode an entire scanline from
the buffer supplied, it 'suspends' itself (i.e. restores its internal
state to the beginning of the last scanline, and returns having consumed
no data).

This means that if we ever have a scanline of jpeg data that is too large
to fit in the stream buffer, we will go into an infinite loop.

This commit is a partial fix for the issue; if we return from the jpeg
lib having consumed no data and having produced no output, we attempt to
compact the data in the jpeg stream. We do this by compacting strings of
0xFF bytes to a single 0xFF byte. (Strings of 0xFF bytes are forbidden by
the spec, but are treated by the library as a single one).

This is sufficient to solve the problem with the supplied corrupted file,
but still leaves us vulnerable to files that have more than 0x800 bytes
in any given single scanline.

A proper fix would require us to spot that the buffer was 'full' and that
the compaction process saved no bytes. Spotting that the compaction
process saved no bytes is easy. Spotting that the buffer is 'full' is not
currently possible as stream->cbsize is not available at the point
where the test needs to be done.