Bug 692481: Fix for remainder of issue; oversize JPEG data causes hang.
If the JPEG lib exits without decoding anything, and the buffer was full,
attempt a compaction; if the compaction succeeds (bytes were saved), then
continue, otherwise bale with an error to avoid us infinitely looping
as we repeatedly try to decode the same data.

The 'buffer is full' condition is the tricky one; we cannot know the
exact size of the buffer, but we can know the minimum size we requested,
so we use that as a close approximation.

Also tweak the pxl handling of error conditions passed back from the
jpeg decoder. Ensure that pxls use of streams set the templates.