Speed up of lcms 16bit tetrahedral interpolation code.
This color table interpolation is used primarily where we have 16 bit image data or
are doing spatial interpolation of image data.  The change introduces
some very minor changes across quite a few files that have 16 bit
data or for which spatial interpolation is occurring.  No changes occur at the white
or black table points.

The code differences involved moving the decision tree outside the loop and performing
the normalization divide by an add and shift.

This change can be readily moved to lcms2.