Bug #692136 Change the FontBBox we use when a type 3 font has a NULL bbox.
PCL creates type 3 fonts where the BBox is 0 0 0 0. This causes Acrobat problems,
so we replace it with a 'default'. Previously this was 0 0 1000 1000, but this
is incorrect for the type 3 fonts created from PCL, and makes it impossible
to search/highlight text in the resulting PDF.

We now check for a type 3 font with an all zero BBox, and instead use 0 0 1 -1
which for PCL fonts works well. Regression tests show no differences with this
change, and hopefully its no more incorrect than it was before.

No differences expected.